+++ jstests/core/do_txn_atomicity.js
@@ -4,6 +4,14 @@
 (function() {
     'use strict';
 
+    // For isMMAPv1.
+    load("jstests/concurrency/fsm_workload_helpers/server_types.js");
+
+    if (isMMAPv1(db)) {
+        jsTestLog("Skipping test as the storage engine does not support doTxn.");
+        return;
+    }
+
     var t = db.doTxn;
     t.drop();
     assert.writeOK(t.insert({_id: 1}));
